criando tratamento de exceções

+++ backend/src/main/java/com/example/to_do_list_backend/service/TarefaService.java
@@ -1,10 +1,11 @@
 package com.example.to_do_list_backend.service;
-
 import com.example.to_do_list_backend.domain.Tarefa;
 import com.example.to_do_list_backend.dto.TarefaRequestDTO;
 import com.example.to_do_list_backend.dto.TarefaResponseDTO;
 import com.example.to_do_list_backend.mapper.TarefaMapper;
 import com.example.to_do_list_backend.repository.TarefaRepository;
+import com.example.to_do_list_backend.service.exception.MissingDataException;
+import com.example.to_do_list_backend.service.exception.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,20 +21,22 @@ public class TarefaService {
     }
 
     public TarefaResponseDTO findById(Long id) {
-        return TarefaMapper.tarefaToResponseDTO(repository.findById(id).orElseThrow());
+        return TarefaMapper.tarefaToResponseDTO(repository.findById(id).orElseThrow(() -> new ResourceNotFoundException(id)));
     }
 
     public TarefaResponseDTO insert(TarefaRequestDTO requestDTO) {
+        checksMissingData(requestDTO);
         return TarefaMapper.tarefaToResponseDTO(repository.save(TarefaMapper.requestDTOToTarefa(requestDTO)));
     }
 
     public void deleteById(Long id) {
-        findById(id);
+        if (!repository.existsById(id)) throw new ResourceNotFoundException(id);
         repository.deleteById(id);
     }
 
     public TarefaResponseDTO update(Long id, TarefaRequestDTO requestDTO) {
-        Tarefa tarefa = repository.findById(id).orElseThrow();
+        checksMissingData(requestDTO);
+        Tarefa tarefa = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException(id));
         updateData(tarefa, requestDTO);
         repository.save(tarefa);
         return TarefaMapper.tarefaToResponseDTO(tarefa);
@@ -44,5 +47,11 @@ public class TarefaService {
         tarefa.setTitle(requestDTO.title());
         tarefa.setDescription(requestDTO.description());
         tarefa.setStatus(requestDTO.status().getCode());
+    }
+
+    //Lança uma exceção caso algum campo não seja preenchido
+    private void checksMissingData(TarefaRequestDTO requestDTO) {
+        if (requestDTO.title() == null || requestDTO.description() == null || requestDTO.status() == null)
+            throw new MissingDataException("Todos os campos devem ser preenchidos.");
     }
 }
